# docs/openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiskOrbit Risk Scoring & Investigation API",
    "description": "Phase 1 + Phase 2 \u2014 Coordinated Refund Abuse Detection & Relationship Investigation. Phase 1: transaction-level risk scoring. Phase 2: bounded graph investigation and grounded evidence generation.",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v2/ops/auth/login": {
      "post": {
        "tags": [
          "Authentication & Session"
        ],
        "summary": "Login",
        "description": "Authenticate user via email, username, or demo user identifier.\nReturns Bearer token and validated user context with role capabilities.",
        "operationId": "login_api_v2_ops_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/auth/session": {
      "get": {
        "tags": [
          "Authentication & Session"
        ],
        "summary": "Validate Session",
        "description": "Validate the current active session token and return identity + permissions.",
        "operationId": "validate_session_api_v2_ops_auth_session_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "analyst_01",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ANALYST",
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "Risk Analyst",
              "title": "X-User-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionValidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/ops/auth/logout": {
      "post": {
        "tags": [
          "Authentication & Session"
        ],
        "summary": "Logout",
        "description": "Invalidate session token and terminate analyst session.",
        "operationId": "logout_api_v2_ops_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "analyst_01",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "ANALYST",
              "title": "X-User-Role"
            }
          },
          {
            "name": "x-user-name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "Risk Analyst",
              "title": "X-User-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username_or_email": {
            "type": "string",
            "title": "Username Or Email"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "remember_me": {
            "type": "boolean",
            "title": "Remember Me",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username_or_email"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserContext"
          },
          "expires_at": {
            "type": "number",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "session_id",
          "user",
          "expires_at"
        ],
        "title": "LoginResponse"
      },
      "LogoutResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session successfully terminated."
          }
        },
        "type": "object",
        "title": "LogoutResponse"
      },
...
      "SessionValidateResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserContext"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "user"
        ],
        "title": "SessionValidateResponse"
      },
...
      "UserContext": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "role",
          "name"
        ],
        "title": "UserContext",
        "description": "Authenticated user context."
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "VIEWER",
          "ANALYST",
          "SENIOR_ANALYST",
          "ADMIN"
        ],
        "title": "UserRole",
        "description": "Supported user roles in RiskOrbit operations."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}